docs: further mentioned IE10+ and EDGE specifics
IE10+ and EDGE don't do substring matching for the users input on the values, so the solution provided by this polypill for that behavior includes manipulating the value itself, which leads to a different developer experience / return value on the value property for those browers. This needed to get reflected within the documentation.