Fix width of list box entries
Before this change, an item of the datalistSelect box did not take the same width as the corresponding input box.
For example, when an animal is selected (using IE9) on the test side https://mfranzke.github.io/datalist-polyfill/demos/ie9/ the selection bar does not extend from the left to the complete right of the input box. Setting the width attribute instead minWidth fixes this issue. Now, the blue selection bar of an item has the same width than the input box.